Break project into multiple files
